// ELSA.Web.Client/src/enums/question.type.ts
export enum QuestionType {
    Essay = 1, // Short text answer question
    Binary = 2, // Yes no question
    MultiChoice = 3, // For multi options question
    SingleChoice = 4, // For single choice question
    ImageSingleChoice = 5, // For multi options question but use image
    ImageMultiChoice = 6, // For single choice question but use image
}